Output KML in UTF-8, without XML character encoding.
Before this change, fields like the description would end up wrongfully
encoded in the KML output. They were encoded using XML encoding (&#123),
which etree treated literally, resulting in (&amp#123).

Now, fastkml/etree simply gets unicode. Due to utf-8 being valid in XML
it does not do anything with it. If that was not the case it's still
recommended to give unicode to etree and let it deal with encoding during
serialization.